OAuth support with permissions is done. Frontend also is uwu. Could improve user selection for FolderClaims. Closes #1

+++ MappedFolderServer/Controllers/SlugApi.cs
@@ -1,3 +1,4 @@
+using MappedFolderServer.Auth;
 using MappedFolderServer.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,24 +10,39 @@ namespace MappedFolderServer.Controllers;
 [Authorize(AuthenticationSchemes = "oidc")]
 public class SlugApi : Controller
 {
-    private readonly IWebHostEnvironment _env;
+    private readonly ICurrentUserService _currentUser;
     private readonly AppDatabaseContext _db;
-    
-    public SlugApi(IWebHostEnvironment env, AppDatabaseContext db)
+
+    public SlugApi(ICurrentUserService currentUser, AppDatabaseContext db)
     {
-        _env = env;
+        _currentUser = currentUser;
         _db = db;
     }
 
     [HttpGet("all")]
     public IActionResult GetAll()
     {
-        return Ok(_db.Mappings.ToList());
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
+        if (loggedInUser.IsAdmin)
+        {
+            return Ok(_db.Mappings.ToList());
+        }
+        return Ok(_db.Mappings.Where(x => x.CreatedBy == loggedInUser.Id).ToList());
     }
 
     [HttpPost]
     public IActionResult Create([FromBody] SlugEntry slugEntry)
     {
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
+        slugEntry.CreatedBy = slugEntry.EditedBy = loggedInUser.Id;
+        slugEntry.FolderPath = Path.GetFullPath(slugEntry.FolderPath);
+        if (!loggedInUser.CanAccessFolder(slugEntry.FolderPath, _db))
+        {
+            return BadRequest(
+                new ApiError("You're not allowed to access this directory. Please ask the admin for permission"));
+        }
         if (!Directory.Exists(slugEntry.FolderPath))
         {
             return BadRequest(new ApiError("Requested directory doesn't exist"));
@@ -45,8 +61,22 @@ public class SlugApi : Controller
     [HttpPost("{id:guid}")]
     public IActionResult Update([FromRoute] Guid id, [FromBody] SlugEntry slugEntry)
     {
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
         SlugEntry? m = _db.Mappings.FirstOrDefault(x => x.Id == id);
         if (m == null) return NotFound();
+        if (!m.CanBeAccessedBy(loggedInUser)) return Forbid();
+        if (_db.Mappings.Any(x => x.Slug == slugEntry.Slug && x.Id != id))
+        {
+            return BadRequest(new ApiError("Slug already exists, please choose a different slug"));
+        }
+        slugEntry.FolderPath = Path.GetFullPath(slugEntry.FolderPath);
+        if (!loggedInUser.CanAccessFolder(slugEntry.FolderPath, _db))
+        {
+            return BadRequest(
+                new ApiError("You're not allowed to access this directory. Please ask the admin for permission"));
+        }
+        m.EditedBy = loggedInUser.Id;
         m.Slug = slugEntry.Slug;
         m.FolderPath = slugEntry.FolderPath;
         m.IsPublic = slugEntry.IsPublic;
@@ -62,8 +92,11 @@ public class SlugApi : Controller
     [HttpDelete("{id:guid}")]
     public IActionResult Delete([FromRoute] Guid id)
     {
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
         SlugEntry? m = _db.Mappings.FirstOrDefault(x => x.Id == id);
         if (m == null) return NotFound();
+        if (!m.CanBeAccessedBy(loggedInUser)) return Forbid();
         _db.Mappings.Remove(m);
         _db.SaveChanges();
         return Ok();
@@ -72,8 +105,11 @@ public class SlugApi : Controller
     [HttpPost("{id:guid}/password")]
     public IActionResult Update([FromRoute] Guid id, [FromBody] LoginRequest request)
     {
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
         SlugEntry? m = _db.Mappings.FirstOrDefault(x => x.Id == id);
         if (m == null) return NotFound();
+        if (!m.CanBeAccessedBy(loggedInUser)) return Forbid();
         m.PasswordSalt = BCrypt.Net.BCrypt.GenerateSalt();
         m.PasswordHash = BCrypt.Net.BCrypt.HashPassword(request.Password, m.PasswordSalt);
         _db.Mappings.Update(m);
@@ -84,8 +120,11 @@ public class SlugApi : Controller
     [HttpDelete("{id:guid}/password")]
     public IActionResult Update([FromRoute] Guid id)
     {
+        User? loggedInUser = _currentUser.GetCurrentUser();
+        if (loggedInUser == null) return Unauthorized();
         SlugEntry? m = _db.Mappings.FirstOrDefault(x => x.Id == id);
         if (m == null) return NotFound();
+        if (!m.CanBeAccessedBy(loggedInUser)) return Forbid();
         m.PasswordHash = null;
         _db.Mappings.Update(m);
         _db.SaveChanges();
